# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Localhost server"
        },
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local server"
        },
        {
            "url": "https://api.example.com"
        }
    ],
    "paths": {
...
        "/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get authenticated user's reservations",
                "operationId": "69fb15b9890a0559eb50e5e1777d9b64",
                "responses": {
                    "200": {
                        "description": "Reservations retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reservations retrieved successfully"
                                        },
                                        "reservations": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No reservations found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Create a new reservation",
                "operationId": "1aabc7d432cc652e63cfc3fb514132cd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hotel_id",
                                    "number_guests",
                                    "room_id",
                                    "check_in",
                                    "check_out"
                                ],
                                "properties": {
                                    "hotel_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "number_guests": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "room_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "check_in": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-01"
                                    },
                                    "check_out": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reservation created successfully"
                                        },
                                        "reservation": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Room not available for selected dates"
                    },
                    "404": {
                        "description": "Error creating reservation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/reservations/{id}": {
            "put": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Update an existing reservation",
                "operationId": "3cb121a7272b76103c947578223bb7d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the reservation to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "room_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "check_in": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-01"
                                    },
                                    "check_out": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reservation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "updated reservation": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Update successful"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reservation not found"
                    },
                    "409": {
                        "description": "Room not available for selected dates"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Cancel an existing reservation",
                "operationId": "fa964cce3285d6c10989429c544421ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the reservation to cancel",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 9
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation cancelled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Reservation cancelled successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reservation not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format: Bearer {token}"
            }
        }
    },
...
}